testsuite: combined cpu and gpu tests.
